ading task 102

+++ 0x05-processes_and_signals/102-zombie.c
@@ -30,7 +30,7 @@ int main(void)
 			exit(0);
 		else
 			printf("Zombie process created, PID: %ld\n",
-					(long) zombie);
+					zombie);
 	}
 	return (infinite_while());
 }
